Extract truthy rewrite into dedicated pass
